requirements.txt
python-dotenv
sqlalchemy
sqlalchemy_utils
psycopg2
requests
transformers==4.26.1
torch
wheel
clip-interrogator==0.5.4
elasticsearch==7.10.1